Move BUYCollectionSort enum into storefront category file.
