+++ Mobile Buy SDK/Mobile Buy SDK/Data/BUYClient+Storefront.h
@@ -32,6 +32,44 @@ NS_ASSUME_NONNULL_BEGIN
 @class BUYProduct;
 
 /**
+ *  The sort order for products in a collection
+ */
+typedef NS_ENUM(NSUInteger, BUYCollectionSort) {
+	/**
+	 *  Sort products by best selling using the order set in the shop's admin
+	 */
+	BUYCollectionSortCollectionDefault,
+	/**
+	 *  Sort products by best selling
+	 */
+	BUYCollectionSortBestSelling,
+	/**
+	 *  Sort products by title, ascending
+	 */
+	BUYCollectionSortTitleAscending,
+	/**
+	 *  Sort products by title, descending
+	 */
+	BUYCollectionSortTitleDescending,
+	/**
+	 *  Sort products by price (first variant), ascending
+	 */
+	BUYCollectionSortPriceAscending,
+	/**
+	 *  Sort products by price (first variant), descending
+	 */
+	BUYCollectionSortPriceDescending,
+	/**
+	 *  Sort products by creation date, ascending
+	 */
+	BUYCollectionSortCreatedAscending,
+	/**
+	 *  Sort products by creation date, descending
+	 */
+	BUYCollectionSortCreatedDescending
+};
+
+/**
  *  Return block containing a BUYShop and/or an NSError
  *
  *  @param shop  A BUYShop object
